Fixed wrong ".dot" usages

+++ mlfromscratch-java/src/main/java/mlfromscratch/deeplearning/layer/optimizers/Optimizer.java
@@ -3,7 +3,7 @@ package mlfromscratch.deeplearning.layer.optimizers;
 import mlfromscratch.math.ndarray.NDArray;
 
 public abstract class Optimizer {
-  public double learning_rate;
+  public float learning_rate;
   public NDArray w_updt;
   
   abstract NDArray update(NDArray w, NDArray gradWrtW);
